# api-docs/api/swagger/swagger.yaml
openapi: 3.0.0
info:
    title: 'Guardian'
    description: 'The Guardian is a modular open-source solution that includes best-in-class identity management and decentralized ledger technology (DLT) libraries. At the heart of the Guardian solution is a sophisticated Policy Workflow Engine (PWE) that enables applications to offer a requirements-based tokenization implementation.'
    version: '1.2.1'
    contact:
        name: 'API developer'
        url: 'https://envisionblockchain.com'
        email: 'info@envisionblockchain.com'
    license:
        name: 'Apache 2.0'
        url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
servers:
    - url: /api/v1
      description: version 1.0
components:
    securitySchemes:
        bearerAuth:
            type: 'http'
            scheme: 'bearer'
            bearerFormat: 'JWT'
    schemas:
        Credentials:
            type: object
            required:
                - username
                - password
            properties:
                username:
                    type: string
                password:
                    type: string
        Account:
            type: object
            required:
                - username
                - role
            properties:
                username:
                    type: string
                role:
                    type: string
                did:
                    type: string
        Session:
            type: object
            required:
                - username
                - role
                - accessToken
            properties:
                username:
                    type: string
                role:
                    type: string
                accessToken:
                    type: string
...
        Error:
            type: object
            required:
                - code
                - message
            properties:
                code:
                    type: number
                message:
                    type: string
...
        HederaAccount:
            type: object
            required:
                - id
                - key
            properties:
                id:
                    type: string
                key:
                    type: string
...
        LogFilters:
            type: object
            properties:
                type:
                    type: string
                startDate:
                    type: string
                endDate:
                    type: string
                attributes:
                    type: array
                    items:
                        type: string
                message:
                    type: string
                pageSize:
                    type: number
                pageIndex:
                    type: number
                sortDirection:
                    type: string
                    enum: [ASC, DESC]
        Log:
            type: object
            properties:
                type:
                    type: string
                datetime:
                    type: string
                message:
                    type: string
                attributes:
                    type: array
                    items:
                        type: string
...
paths:
    /accounts:
        get:
            tags:
                - accounts
            description: Returns all users except those with roles Root Authority and Auditor. Only users with the Root Authority role are allowed to make the request.
            security:
                - bearerAuth: []
            summary: Returns a list of users, excluding Root Authority and Auditors.
            responses:
                200:
                    description: Successful operation.
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/Account'
                401:
                    description: Unauthorized.
                403:
                    description: Forbidden.
                500:
                    description: Internal server error.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
    /accounts/session:
        get:
            tags:
                - accounts
            description: Returns current user session.
            security:
                - bearerAuth: []
            summary: Returns current session of the user.
            responses:
                200:
                    description: Successful operation.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Session'
                401:
                    description: Unauthorized.
                500:
                    description: Internal server error.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
    /accounts/register:
        post:
            tags:
                - accounts
            summary: Registers a new user account.
            requestBody:
                description: Object that contain username, password and role (optional) fields.
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Credentials'
            responses:
                201:
                    description: Successful operation.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Account'
                500:
                    description: Internal server error.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
    /accounts/login:
        post:
            tags:
                - accounts
            summary: Logs user into the system.
            requestBody:
                description: Object that contains username and password fields.
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Credentials'
            responses:
                200:
                    description: Successful operation.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Session'
                500:
                    description: Internal server error.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
...
    /demo/registeredUsers:
        get:
            tags:
                - demo
            deprecated: true
            description: Returns all registered user accounts.
            responses:
                200:
                    description: All registered user accounts.
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/Account'
                500:
                    description: Internal server error.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
    /demo/randomKey:
        get:
            tags:
                - demo
            deprecated: true
            description: Generates a new Hedera account with a random private key.
            security:
                - bearerAuth: []
            responses:
                200:
                    description: Successful operation.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/HederaAccount'
                401:
                    description: Unauthorized.
                403:
                    description: Forbidden.
                500:
                    description: Internal server error.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
...
    /logs:
        post:
            tags:
                - logs
            description: Returns logs. For users with the Root Authority role only.
            security:
                - bearerAuth: []
            requestBody:
                description: Log filters.
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/LogFilters'
            summary: Returns logs.
            responses:
                200:
                    description: Successful operation.
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    totalCount:
                                        type: number
                                    logs:
                                        $ref: '#/components/schemas/Log'
                401:
                    description: Unauthorized.
                403:
                    description: Forbidden.
                500:
                    description: Internal server error.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
    /logs/attributes:
        get:
            tags:
                - logs
            description: Returns logs attributes. For users with the Root Authority role only.
            parameters:
                - in: query
                  name: name
                  schema:
                      type: string
                  description: Part of name.
            security:
                - bearerAuth: []
            summary: Returns logs attributes.
            responses:
                200:
                    description: Successful operation.
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: string
                401:
                    description: Unauthorized.
                403:
                    description: Forbidden.
                500:
                    description: Internal server error.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'